Increases numeric precision of sound effect 'length in milliseconds' field (from long to double) to prevent rounding issues when converting this length into audio frames. Fixes #82.
Also, some cleanup with seconds/milliseconds.

The length in milliseconds of the double whistle sound effect is around 999.8 ms, but this was previously represented as a long field which got rounded downwards to 999 ms. When using this to calculate the audio frames we ended up with an odd number of frames which, after being inserted into the output audio stream, borked the byte order and looked like white noise in the player.